convert marketplace activity to fragment + navigation
