+++ app/build.gradle
@@ -58,4 +58,9 @@ dependencies {
 
     // LiveData + ViewModel
     implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
+
+    // Navigation
+    implementation "androidx.navigation:navigation-fragment-ktx:$version_navigation"
+    implementation "androidx.navigation:navigation-ui-ktx:$version_navigation"
+    implementation "androidx.legacy:legacy-support-v4:1.0.0"
 }
