+++ app/src/main/java/com/example/mystable/screens/MainActivity.kt
@@ -1,0 +1,19 @@
+package com.example.mystable.screens
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.databinding.DataBindingUtil
+import com.example.mystable.R
+import com.example.mystable.databinding.ActivityMarketplaceBinding
+
+class MainActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityMarketplaceBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        binding = DataBindingUtil.setContentView(this, R.layout.activity_marketplace)
+
+
+    }
+}
